api/flash: Clear the EEPROM after backing it up
In case we flash firmware that changes the Storage layout, we do not want to end
up in a situation where an area that was previously used by one plugin, gets
assigned to another, and have random data in it.

To avoid this situation, we have to clear EEPROM during flashing, right after
backing it up.

Fixes #934.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>